add schema test to check that household_id and simulant_id stick around
